requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
requests==2.28.2
six==1.16.0
urllib3==1.26.15
Werkzeug==2.2.3
